requirements.txt
pyspark
pandas
numpy 
pandas 
polars 
openpyxl
tqdm
sqlalchemy
python-dotenv
fastparquet
fastapi[all]
packaging